Slightly faster 7-segment display refresh function
Rather than storing characters in the buffer, and converting the characters into patterns 280 times a second, this commit changes the buffer into a buffer of patterns. The characters are converted into patterns only once at the buffer write function.

This change should be safe in terms of IRQ/interrupt considerations, but please double-check me on that.